put helper function in utils
